Fix: Move classes into separate files
