+++ tests/fixture/DeleteRoutingChainTestCommand.php
@@ -1,0 +1,11 @@
+<?php declare(strict_types=1);
+
+namespace Lucille\UnitTests;
+
+use Lucille\Command;
+use Lucille\Result\Result;
+
+class DeleteRoutingChainTestCommand implements Command {
+    public function execute(): Result {
+    }
+}
