+++ tests/fixture/GetRoutingChainTestQuery.php
@@ -1,0 +1,11 @@
+<?php declare(strict_types=1);
+
+namespace Lucille\UnitTests;
+
+use Lucille\Query;
+use Lucille\Result\Result;
+
+class GetRoutingChainTestQuery implements Query {
+    public function execute(): Result {
+    }
+}
