+++ tests/fixture/InvalidRoutingChain.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Lucille\UnitTests;
+
+use Lucille\Routing\RoutingChain;
+
+class InvalidRoutingChain implements RoutingChain {
+}
