+++ tests/fixture/PatchRoutingChainTestReturnsCommandRouter.php
@@ -1,0 +1,13 @@
+<?php declare(strict_types=1);
+
+namespace Lucille\UnitTests;
+
+use Lucille\Command;
+use Lucille\Request\PatchRequest;
+use Lucille\Routing\PatchRouter;
+
+class PatchRoutingChainTestReturnsCommandRouter extends PatchRouter {
+    public function route(PatchRequest $request): Command {
+        return new PatchRoutingChainTestCommand();
+    }
+}
