+++ tests/fixture/PutRoutingChainTestCommand.php
@@ -1,0 +1,11 @@
+<?php declare(strict_types=1);
+
+namespace Lucille\UnitTests;
+
+use Lucille\Command;
+use Lucille\Result\Result;
+
+class PutRoutingChainTestCommand implements Command {
+    public function execute(): Result {
+    }
+}
