+++ tests/fixture/ResultRoutingChainTestResponse.php
@@ -1,0 +1,11 @@
+<?php declare(strict_types=1);
+
+namespace Lucille\UnitTests;
+
+use Lucille\Response\Response;
+
+class ResultRoutingChainTestResponse implements Response {
+    public function send(): void {
+        die('dump');
+    }
+}
