+++ tests/fixture/TestEx.php
@@ -1,0 +1,6 @@
+<?php declare(strict_types=1);
+
+namespace Lucille\UnitTests;
+
+class TestEx extends \Exception {
+}
