+++ tests/fixture/TestResponse.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types=1);
+
+namespace Lucille\UnitTests;
+
+use Lucille\Response\GenericResponse;
+
+class TestResponse extends GenericResponse {
+    public function send(): void {
+    }
+}
