+++ tests/fixture/TestResult.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Lucille\UnitTests;
+
+use Lucille\Result\Result;
+
+class TestResult implements Result {
+}
