+++ tests/framework/routing/chain/DeleteRoutingChainTest.php
@@ -16,29 +16,10 @@ use Lucille\Request\Body\EmptyRequestBody;
 use Lucille\Request\DeleteRequest;
 use Lucille\Request\Parameter\RequestParameterCollection;
 use Lucille\Request\Uri;
-use Lucille\Result\Result;
 use Lucille\Routing\DeleteRouter;
 use Lucille\Routing\DeleteRoutingChain;
 
 use PHPUnit\Framework\TestCase;
-
-class DeleteRoutingChainTestCommand implements Command {
-    public function execute(): Result {
-    }
-}
-
-class DeleteRoutingChainTestRouter extends DeleteRouter {
-    public function route(DeleteRequest $request): Command {
-        return $this->getNext()->route($request);
-    }
-}
-
-class DeleteRoutingChainTestReturnsCommandRouter extends DeleteRouter {
-    public function route(DeleteRequest $request): Command {
-        return new DeleteRoutingChainTestCommand();
-    }
-}
-
 
 /**
  * @coversDefaultClass \Lucille\Routing\DeleteRoutingChain
