+++ tests/framework/routing/router/ResultRouterTest.php
@@ -10,16 +10,8 @@
 namespace Lucille\UnitTests;
 
 use Lucille\Exceptions\RoutingChainConfigurationException;
-use Lucille\Response\Response;
-use Lucille\Result\Result;
 use Lucille\Routing\ResultRouter;
 use PHPUnit\Framework\TestCase;
-    
-class ResultTestRouter extends ResultRouter {
-    public function route(Result $request): Response {
-        return $this->getNext()->route($request);
-    }
-}
 
 /**
  * @coversDefaultClass \Lucille\Routing\ResultRouter
